--- src/features/users/routes/Profile/TabMenu/BookmarkList.style.ts
@@ -1,8 +1,0 @@
-import styled from "@emotion/styled";
-
-export const Target = styled.div`
-  position: relative;
-  top: -15px;
-  background-color: transparent;
-  height: 15px;
-`;
