+++ src/features/users/routes/Profile/TabMenu/index.tsx
@@ -1,10 +1,16 @@
-import { useEffect, useState } from "react";
+import { useInfiniteQuery } from "@tanstack/react-query";
+import { useEffect, useRef, useState } from "react";
 import { useLocation } from "react-router-dom";
+
+import useAuth from "@/features/auth/hooks/useAuth";
+import useSortBar from "@/features/users/hooks/useSortBar";
+import { useApi } from "@/hooks/useApi";
+import useIntersectionObserver from "@/hooks/useIntersectionObserver";
 
 import BookmarkList from "./BookmarkList";
 import ReviewList from "./ReviewList";
 import SortBar from "./SortBar";
-import { ContentContainer, Tab, TabButton } from "./style";
+import { ContentContainer, Tab, TabButton, Target } from "./style";
 
 const REVIEW_MENU = "한줄리뷰";
 const BOOKMARK_MENU = "입덕애니";
@@ -21,8 +27,31 @@ interface TabMenuProps {
 
 export default function TabMenu({ isMine }: TabMenuProps) {
   const location = useLocation();
+  const targetRef = useRef(null);
   const [selectedMenu, setSelectedMenu] = useState<MENU>(REVIEW_MENU);
   const handleTabMenuClick = (text: MENU) => setSelectedMenu(text);
+  const { selected, SHEET_BUTTONS, handleSortClick } = useSortBar(selectedMenu);
+  const {
+    user: { memberId },
+  } = useAuth();
+  const { profile } = useApi();
+  const { data, isLoading, fetchNextPage, hasNextPage } = useInfiniteQuery(
+    ["bookmarks", memberId],
+    ({ pageParam }) => profile.getBookmark(memberId, pageParam),
+    {
+      getNextPageParam: (lastPage) => lastPage.lastId || undefined,
+      select: (data) => ({
+        pages: data.pages.flatMap((page) => page.items),
+        pageParams: data.pageParams,
+      }),
+    },
+  );
+
+  useIntersectionObserver({
+    target: targetRef,
+    onIntersect: fetchNextPage,
+    enabled: hasNextPage,
+  });
 
   useEffect(() => {
     if (!location.state) return;
@@ -45,9 +74,20 @@ export default function TabMenu({ isMine }: TabMenuProps) {
         ))}
       </Tab>
       <ContentContainer>
-        <SortBar menu={selectedMenu} />
+        <SortBar
+          count={data?.pages.length ?? 0}
+          selected={selected}
+          BUTTONS={SHEET_BUTTONS}
+          onClick={handleSortClick}
+        />
         {selectedMenu === "한줄리뷰" && <ReviewList isMine={isMine} />}
-        {selectedMenu === "입덕애니" && <BookmarkList />}
+        {selectedMenu === "입덕애니" && (
+          <BookmarkList list={data?.pages ?? []} />
+        )}
+
+        <Target ref={targetRef} />
+
+        {isLoading && <span>로딩중</span>}
       </ContentContainer>
     </>
   );
