+++ src/features/users/routes/Profile/TabMenu/style.ts
@@ -32,3 +32,10 @@ export const ContentContainer = styled.div`
   padding: 0 16px;
   border-top: 2px solid ${({ theme }) => theme.colors.neutral["05"]};
 `;
+
+export const Target = styled.div`
+  position: relative;
+  top: -15px;
+  background-color: transparent;
+  height: 15px;
+`;
